requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
google-adk
google-genai
